Rename Rabbit to daMip_c and promote ov085/daMip_c into one C++ TU
Two things that cannot be split: the cartridge's own name for the class,
and the fold of its 32 shards into src/actors/daMip_c.cpp.  dsd resolves
the sixteen pointer-to-member records in ov085 .data by symbol NAME, so
a half-landed rename links every one of them as 0x00000000.

THE NAME IS READ, NOT COINED.  ov085 0x021300bc holds the nine bytes
37 64 61 4d 69 70 5f 63 00 -- the length-prefixed string "7daMip_c" --
and 0x021300c8 is the __si_class_type_info record that names it, whose
base word resolves to _ZTI12dEnemyBase_c in ov002.  `Rabbit` was a
coined tree name and is spelled nowhere in the image.  The rename costs
eight already-mangled _ZN6Rabbit* rows, and they move here together with
config/arm9/overlays/ov085/symbols.txt, include/Rabbit.h -> daMip_c.h and
nine ledger rows recording the old spelling.

  tubuild.py verify   32/32 MATCH, objisolate clean, reloc-destinations
                      clean -> TEXT-VERIFIED; emission order all 32 in
                      ROM-ascending section order
  tubuild.py linkcheck --baseline over the promoted tree: 9136 enrolled
                      sources, 0 compile failures, 10935/10935 source-built
                      functions reproducing, module fidelity 106/106 exact
                      at 100.000000% of compared bytes, ROM sha256
                      d1506e90efae5e2d2cf119926a4ac2a291bd5ca78349d09d5024e1a918c478e8

The per-TU `linkcheck ov085/daMip_c` cannot run on this class and the
manifest records why.  The promotion carries a symbol rename, so no
intermediate tree links: before it the merged TU defines names
symbols.txt does not carry, after it the 32 shards define names it no
longer carries.  And once the spliced delinks.txt entry exists,
tubuild.py:4641 routes the id through tu_production.prepare_intact_object,
which refuses a .text-only claim.  The `--baseline` run above is the
available link evidence.  A cleaner sequencing for the next writer is two
commits: fold first under the func_ov085_* names, rename second.

All 32 members convert.  23 shards that were still auto-named
func_ov085_* are now real daMip_c:: methods, and daMip_c_classInit is the
only free function left, correctly -- it is the registry factory.
Member-ness is ROM-proven for SetState and the sixteen State* pair: each
is the target of an 8-byte {fnptr, 0} pointer-to-member constant in
0x0213003c..0x021300bc, which can point at nothing but a member of this
class.  The other six helper names are coined and say so in the header
and in symbols/actor_renames.tsv.

ONE BYTE DIFFERENCE IN THE WHOLE FOLD, and it was a struct spelling.
StateFleeMain copies the player's position out of +0x5c; mwccarm
scalarises a copy of a three-scalar-member struct into three ldr/str
pairs, and the ROM does ldm/stm.  The same twelve bytes spelled as an
array member -- struct V3Blk { s32 w[3]; } -- makes the compiler emit the
block move.  Eight bytes, and the only diff after the merge.

Declarations were reconciled by USE, not by majority.  The 32 legacy
files disagree about ~80 externals; a class member function may not sit
inside a linkage specification, so every call needs one file-scope
extern "C" region with one signature each.  Sound::PlaySub and
_Z14ApproachLinearRsss must return a value (StateTalkMain tests them in
an `if`, so the `void` three of four files declared cannot be right) and
RandomIntInternal is unsigned (InitResources shifts its result right by
8 and the ROM uses LSR).  DATA declarations were NOT reconciled: mwcc
leaves a file-scope variable's name unmangled in C++, so each stays
block-scope in the member that recovered it -- 0x021305c0 alone was
spelled int[], struct G { int w[2]; }, an incomplete struct S, void *[]
and plain char.  decl_common.h is deliberately not included: five of
these members are declared there with void* where the shards define
char*, which is not a legal overload.

dActor_c::ClosestPlayer is called as a real member and that is
load-bearing.  func_ov085_0212b8dc called _ZN8dActor_c13ClosestPlayerEv
with NO argument -- the ROM emits no `mov r0` there -- which no
void*-parameter declaration can reproduce.  The member call passes `this`
implicitly and matches.

compiler_only_output is 11 data rows + 2 functions: 2x(4 ancestors +
self) typeinfo/name records + _ZTV7daMip_c + the homeless
_ZN7daMip_cD2Ev (ov085 runs D1 at 0x0212a6d4 straight into D0 and then
into TestWaterBelow, with no third variant in the image) +
_ZN7Vector3D1Ev, whose canonical copy is arm9:0x020072c0.
romdata_check.check_object: _ZTV7daMip_c VERIFIED, ROM extent 124 bytes
from the 0x021300f8 address point (0x84 storage less the eight-byte ABI
preamble), 0 blind words; 0 DIFFERS across all eleven records.

tiers_ratchet banked +2 CONVERTED (2696 -> 2698): five whole-file
identities out, seven member identities in.  The five backslide rows are
the usual arithmetic of a reconstructed TU -- it must spell vague-linkage
and cross-module symbols directly or the range will not link.
