+++ config/arm9/overlays/ov085/delinks.txt
@@ -163,133 +163,9 @@ src/d_a_peach.c:
     complete
     .text start:0x0212a684 end:0x0212a6d4
 
-src/_ZN6RabbitD1Ev.cpp:
+src/actors/daMip_c.cpp:
     complete
-    .text start:0x0212a6d4 end:0x0212a724
-
-src/_ZN6RabbitD0Ev.cpp:
-    complete
-    .text start:0x0212a724 end:0x0212a788
-
-src/func_ov085_0212a788.c:
-    complete
-    .text start:0x0212a788 end:0x0212a828
-
-src/func_ov085_0212a828.cpp:
-    complete
-    .text start:0x0212a828 end:0x0212a904
-
-src/func_ov085_0212a904.c:
-    complete
-    .text start:0x0212a904 end:0x0212aaa4
-
-src/func_ov085_0212aaa4.c:
-    complete
-    .text start:0x0212aaa4 end:0x0212aaec
-
-src/func_ov085_0212aaec.c:
-    complete
-    .text start:0x0212aaec end:0x0212ac3c
-
-src/func_ov085_0212ac3c.c:
-    complete
-    .text start:0x0212ac3c end:0x0212ac4c
-
-src/func_ov085_0212ac4c.c:
-    complete
-    .text start:0x0212ac4c end:0x0212ad8c
-
-src/func_ov085_0212ad8c.c:
-    complete
-    .text start:0x0212ad8c end:0x0212ae08
-
-src/func_ov085_0212ae08.c:
-    complete
-    .text start:0x0212ae08 end:0x0212b3fc
-
-src/func_ov085_0212b3fc.c:
-    complete
-    .text start:0x0212b3fc end:0x0212b444
-
-src/func_ov085_0212b444.c:
-    complete
-    .text start:0x0212b444 end:0x0212b478
-
-src/func_ov085_0212b478.c:
-    complete
-    .text start:0x0212b478 end:0x0212b4b4
-
-src/func_ov085_0212b4b4.c:
-    complete
-    .text start:0x0212b4b4 end:0x0212b75c
-
-src/func_ov085_0212b75c.cpp:
-    complete
-    .text start:0x0212b75c end:0x0212b86c
-
-src/func_ov085_0212b86c.c:
-    complete
-    .text start:0x0212b86c end:0x0212b8a0
-
-src/func_ov085_0212b8a0.c:
-    complete
-    .text start:0x0212b8a0 end:0x0212b8dc
-
-src/func_ov085_0212b8dc.cpp:
-    complete
-    .text start:0x0212b8dc end:0x0212bc14
-
-src/func_ov085_0212bc14.c:
-    complete
-    .text start:0x0212bc14 end:0x0212bc78
-
-src/func_ov085_0212bc78.cpp:
-    complete
-    .text start:0x0212bc78 end:0x0212bcc8
-
-src/func_ov085_0212bcc8.c:
-    complete
-    .text start:0x0212bcc8 end:0x0212bdbc
-
-src/func_ov085_0212bdbc.cpp:
-    complete
-    .text start:0x0212bdbc end:0x0212bedc
-
-src/func_ov085_0212bedc.cpp:
-    complete
-    .text start:0x0212bedc end:0x0212c004
-
-src/_ZN6Rabbit16CleanupResourcesEv.cpp:
-    complete
-    .text start:0x0212c004 end:0x0212c070
-
-src/_ZN6Rabbit16OnPendingDestroyEv.cpp:
-    complete
-    .text start:0x0212c070 end:0x0212c074
-
-src/_ZN6Rabbit6RenderEv.cpp:
-    complete
-    .text start:0x0212c074 end:0x0212c150
-
-src/func_ov085_0212c150.cpp:
-    complete
-    .text start:0x0212c150 end:0x0212c230
-
-src/_ZN6Rabbit8BehaviorEv.cpp:
-    complete
-    .text start:0x0212c230 end:0x0212c7fc
-
-src/_ZN6Rabbit13InitResourcesEv.cpp:
-    complete
-    .text start:0x0212c7fc end:0x0212cc18
-
-src/_ZN6Rabbit13OnYoshiTryEatEv.cpp:
-    complete
-    .text start:0x0212cc18 end:0x0212cc2c
-
-src/d_a_mip.c:
-    complete
-    .text start:0x0212cc2c end:0x0212cc88
+    .text start:0x0212a6d4 end:0x0212cc88
 
 src/_ZN9RabbitKeyD1Ev.cpp:
     complete
