+++ config/arm9/overlays/ov085/symbols.txt
@@ -38,37 +38,37 @@ _ZN13PrincessPeach6RenderEv kind:function(arm,size=0x24) addr:0x0212a508
 _ZN13PrincessPeach8BehaviorEv kind:function(arm,size=0x5c) addr:0x0212a52c
 _ZN13PrincessPeach13InitResourcesEv kind:function(arm,size=0xfc) addr:0x0212a588
 daPeach_c_classInit kind:function(arm,size=0x50) addr:0x0212a684
-_ZN6RabbitD1Ev kind:function(arm,size=0x50) addr:0x0212a6d4
-_ZN6RabbitD0Ev kind:function(arm,size=0x64) addr:0x0212a724
-func_ov085_0212a788 kind:function(arm,size=0xa0) addr:0x0212a788
-func_ov085_0212a828 kind:function(arm,size=0xdc) addr:0x0212a828
-func_ov085_0212a904 kind:function(arm,size=0x1a0) addr:0x0212a904
-func_ov085_0212aaa4 kind:function(arm,size=0x48) addr:0x0212aaa4
-func_ov085_0212aaec kind:function(arm,size=0x150) addr:0x0212aaec
-func_ov085_0212ac3c kind:function(arm,size=0x10) addr:0x0212ac3c
-func_ov085_0212ac4c kind:function(arm,size=0x140) addr:0x0212ac4c
-func_ov085_0212ad8c kind:function(arm,size=0x7c) addr:0x0212ad8c
-func_ov085_0212ae08 kind:function(arm,size=0x5f4) addr:0x0212ae08
-func_ov085_0212b3fc kind:function(arm,size=0x48) addr:0x0212b3fc
-func_ov085_0212b444 kind:function(arm,size=0x34) addr:0x0212b444
-func_ov085_0212b478 kind:function(arm,size=0x3c) addr:0x0212b478
-func_ov085_0212b4b4 kind:function(arm,size=0x2a8) addr:0x0212b4b4
-func_ov085_0212b75c kind:function(arm,size=0x110) addr:0x0212b75c
-func_ov085_0212b86c kind:function(arm,size=0x34) addr:0x0212b86c
-func_ov085_0212b8a0 kind:function(arm,size=0x3c) addr:0x0212b8a0
-func_ov085_0212b8dc kind:function(arm,size=0x338) addr:0x0212b8dc
-func_ov085_0212bc14 kind:function(arm,size=0x64) addr:0x0212bc14
-func_ov085_0212bc78 kind:function(arm,size=0x50) addr:0x0212bc78
-func_ov085_0212bcc8 kind:function(arm,size=0xf4) addr:0x0212bcc8
-func_ov085_0212bdbc kind:function(arm,size=0x120) addr:0x0212bdbc
-func_ov085_0212bedc kind:function(arm,size=0x128) addr:0x0212bedc
-_ZN6Rabbit16CleanupResourcesEv kind:function(arm,size=0x6c) addr:0x0212c004
-_ZN6Rabbit16OnPendingDestroyEv kind:function(arm,size=0x4) addr:0x0212c070
-_ZN6Rabbit6RenderEv kind:function(arm,size=0xdc) addr:0x0212c074
-func_ov085_0212c150 kind:function(arm,size=0xe0) addr:0x0212c150
-_ZN6Rabbit8BehaviorEv kind:function(arm,size=0x5cc) addr:0x0212c230
-_ZN6Rabbit13InitResourcesEv kind:function(arm,size=0x41c) addr:0x0212c7fc
-_ZN6Rabbit13OnYoshiTryEatEv kind:function(arm,size=0x14) addr:0x0212cc18
+_ZN7daMip_cD1Ev kind:function(arm,size=0x50) addr:0x0212a6d4
+_ZN7daMip_cD0Ev kind:function(arm,size=0x64) addr:0x0212a724
+_ZN7daMip_c14TestWaterBelowEv kind:function(arm,size=0xa0) addr:0x0212a788
+_ZN7daMip_c10UpdateGrabEv kind:function(arm,size=0xdc) addr:0x0212a828
+_ZN7daMip_c17StateSaveTalkMainEv kind:function(arm,size=0x1a0) addr:0x0212a904
+_ZN7daMip_c17StateSaveTalkInitEv kind:function(arm,size=0x48) addr:0x0212aaa4
+_ZN7daMip_c13StateTalkMainEv kind:function(arm,size=0x150) addr:0x0212aaec
+_ZN7daMip_c13StateTalkInitEv kind:function(arm,size=0x10) addr:0x0212ac3c
+_ZN7daMip_c17StateReleasedMainEv kind:function(arm,size=0x140) addr:0x0212ac4c
+_ZN7daMip_c17StateReleasedInitEv kind:function(arm,size=0x7c) addr:0x0212ad8c
+_ZN7daMip_c15StateCaughtMainEv kind:function(arm,size=0x5f4) addr:0x0212ae08
+_ZN7daMip_c15StateCaughtInitEv kind:function(arm,size=0x48) addr:0x0212b3fc
+_ZN7daMip_c13StateRestMainEv kind:function(arm,size=0x34) addr:0x0212b444
+_ZN7daMip_c13StateRestInitEv kind:function(arm,size=0x3c) addr:0x0212b478
+_ZN7daMip_c13StateFleeMainEv kind:function(arm,size=0x2a8) addr:0x0212b4b4
+_ZN7daMip_c13StateFleeInitEv kind:function(arm,size=0x110) addr:0x0212b75c
+_ZN7daMip_c16StateStartleMainEv kind:function(arm,size=0x34) addr:0x0212b86c
+_ZN7daMip_c16StateStartleInitEv kind:function(arm,size=0x3c) addr:0x0212b8a0
+_ZN7daMip_c13StateIdleMainEv kind:function(arm,size=0x338) addr:0x0212b8dc
+_ZN7daMip_c13StateIdleInitEv kind:function(arm,size=0x64) addr:0x0212bc14
+_ZN7daMip_c8SetStateEPv kind:function(arm,size=0x50) addr:0x0212bc78
+_ZN7daMip_c21UpdateMatrixAndShadowEv kind:function(arm,size=0xf4) addr:0x0212bcc8
+_ZN7daMip_c19UpdateCarriedMatrixEv kind:function(arm,size=0x120) addr:0x0212bdbc
+_ZN7daMip_c18UpdateMirrorShadowEv kind:function(arm,size=0x128) addr:0x0212bedc
+_ZN7daMip_c16CleanupResourcesEv kind:function(arm,size=0x6c) addr:0x0212c004
+_ZN7daMip_c16OnPendingDestroyEv kind:function(arm,size=0x4) addr:0x0212c070
+_ZN7daMip_c6RenderEv kind:function(arm,size=0xdc) addr:0x0212c074
+_ZN7daMip_c17RenderMirrorImageEv kind:function(arm,size=0xe0) addr:0x0212c150
+_ZN7daMip_c8BehaviorEv kind:function(arm,size=0x5cc) addr:0x0212c230
+_ZN7daMip_c13InitResourcesEv kind:function(arm,size=0x41c) addr:0x0212c7fc
+_ZN7daMip_c13OnYoshiTryEatEv kind:function(arm,size=0x14) addr:0x0212cc18
 daMip_c_classInit kind:function(arm,size=0x5c) addr:0x0212cc2c
 _ZN9RabbitKeyD1Ev kind:function(arm,size=0x38) addr:0x0212cc88
 _ZN9RabbitKeyD0Ev kind:function(arm,size=0x4c) addr:0x0212ccc0
@@ -187,7 +187,6 @@ data_ov085_021300b4 kind:data(any) addr:0x021300b4
 _ZTS7daMip_c kind:data(any) addr:0x021300bc
 _ZTI7daMip_c kind:data(any) addr:0x021300c8
 g_profile_MIP kind:data(any) addr:0x021300d4
-_ZTV6Rabbit kind:data(any) addr:0x021300f8
 _ZTV7daMip_c kind:data(any) addr:0x021300f8
 data_ov085_02130174 kind:data(any) addr:0x02130174
 data_ov085_0213017c kind:data(any) addr:0x0213017c
