+++ config/converted-baseline.json
@@ -7,7 +7,7 @@
     "no_codegen_trick",
     "no_mangled_refs"
   ],
-  "count": 2696,
+  "count": 2698,
   "converted": [
     "src/ARMMathLoadState.c",
     "src/ARMMathSaveState.c",
@@ -1393,11 +1393,6 @@
     "src/_ZN6Player7TryGrabER8dActor_c.cpp",
     "src/_ZN6PlayerD0Ev.cpp",
     "src/_ZN6PlayerD1Ev.cpp",
-    "src/_ZN6Rabbit13OnYoshiTryEatEv.cpp",
-    "src/_ZN6Rabbit16CleanupResourcesEv.cpp",
-    "src/_ZN6Rabbit16OnPendingDestroyEv.cpp",
-    "src/_ZN6RabbitD0Ev.cpp",
-    "src/_ZN6RabbitD1Ev.cpp",
     "src/_ZN6ShipUp16CleanupResourcesEv.cpp",
     "src/_ZN6ShipUp6RenderEv.cpp",
     "src/_ZN6ShipUpD0Ev.cpp",
@@ -2164,6 +2159,13 @@
     "src/actors/daDkk_c.cpp#_ZN7daDkk_cD0Ev",
     "src/actors/daDkk_c.cpp#_ZN7daDkk_cD1Ev",
     "src/actors/daDsnBase_c.cpp#_ZN11daDsnBase_c16CleanupResourcesEv",
+    "src/actors/daMip_c.cpp#_ZN7daMip_c13OnYoshiTryEatEv",
+    "src/actors/daMip_c.cpp#_ZN7daMip_c13StateTalkInitEv",
+    "src/actors/daMip_c.cpp#_ZN7daMip_c16CleanupResourcesEv",
+    "src/actors/daMip_c.cpp#_ZN7daMip_c16OnPendingDestroyEv",
+    "src/actors/daMip_c.cpp#_ZN7daMip_c8SetStateEPv",
+    "src/actors/daMip_c.cpp#_ZN7daMip_cD0Ev",
+    "src/actors/daMip_c.cpp#_ZN7daMip_cD1Ev",
     "src/actors/daObjCtMecha03_c.cpp#_ZN16daObjCtMecha03_c6RenderEv",
     "src/actors/daObjCtMecha03_c.cpp#_ZN16daObjCtMecha03_cD0Ev",
     "src/actors/daObjCtMecha03_c.cpp#_ZN16daObjCtMecha03_cD1Ev",
