+++ notes/enemy-leaf-provenance.md
@@ -6,7 +6,7 @@ back, the field stays `unk_NNN` and the reason is recorded — a name nobody can
 check is a claim the next reader would trust for nothing.
 
 Classes covered here: JetStream, Goomboss, daWanwan_c, BobOmb, Whomp,
-RollingIronBall, KoopaShell, Klepto, Rabbit, ChiefChilly.
+RollingIronBall, KoopaShell, Klepto, daMip_c, ChiefChilly.
 
 A recurring source of `unk_` in these headers is the `#else` C twin, which
 restates the whole `fBase_c -> dBase_c -> dActor_c -> dEnemyBase_c` chain flat.
