--- src/_ZN6Rabbit16OnPendingDestroyEv.cpp
@@ -1,12 +1,0 @@
-//cpp
-// @symbol _ZN6Rabbit16OnPendingDestroyEv
-/* recovered: shared header, real C++ method
- *
- * Empty -- the ROM body is a single `bx lr`. The override exists to suppress
- * whatever the base does on pending destroy, not to do anything itself.
- */
-#include "Rabbit.h"
-
-void Rabbit::OnPendingDestroy()
-{
-}
