--- src/func_ov085_0212ac3c.c
@@ -1,4 +1,0 @@
-int func_ov085_0212ac3c(int *p)
-{
-    p[263] = 0; return 1;
-}
