--- src/func_ov085_0212b8a0.c
@@ -1,6 +1,0 @@
-extern int _ZN9ModelAnim7SetAnimEP8BCA_Filei5Fix12IiEj();
-extern int *data_ov085_021305c8[];
-int func_ov085_0212b8a0(char *c){
-  _ZN9ModelAnim7SetAnimEP8BCA_Filei5Fix12IiEj((char*)c+0x300, data_ov085_021305c8[1], 0x40000000, 0x1000, 0);
-  return 1;
-}
