--- src/func_ov085_0212bedc.cpp
@@ -1,67 +1,0 @@
-//cpp
-// @symbol func_ov085_0212bedc
-/* recovered: shared common types */
-#include "common.h"
-typedef int Fix12;
-struct Vector3_local { int x; int y; int z; };
-
-
-extern "C" int Vec3_HorzDist(const Vector3_local *a, const Vector3_local *b);
-extern "C" short Vec3_HorzAngle(const Vector3_local *v0, const Vector3_local *v1);
-extern "C" void Matrix4x3_FromRotationY(void *m, int angle);
-extern "C" void MulVec3Mat4x3(const Vector3_local *v, const Matrix4x3 *m, Vector3_local *res);
-extern "C" void Matrix4x3_FromTranslation(Matrix4x3 *m, int x, int y, int z);
-
-struct ShadowModel;
-struct Matrix4x3;
-
-struct dActor_c {
-    void DropShadowRadHeight(ShadowModel &, Matrix4x3 &, Fix12, Fix12, unsigned int);
-};
-/* Signature deliberately copied from the local declaration above: the
-   ROM name carries by-value class parameters (e.g. Fix12<int>), which
-   mwccarm passes differently at the call site, so declaring the true
-   types breaks the byte match. See notes/mwccarm-codegen.md 6az. */
-extern "C" void _ZN8dActor_c19DropShadowRadHeightER11ShadowModelR9Matrix4x35Fix12IiES5_j(void *, ShadowModel &, Matrix4x3 &, Fix12, Fix12, unsigned int);
-
-
-extern Matrix4x3 data_020a0e68;
-
-struct Obj {
-    char pad5c[0x5c];
-    int v5c;
-    int v60;
-    int v64;
-    char pad3c0[0x3c0 - 0x68];
-    char shadowmodel[0x28];
-    Matrix4x3 mtx;
-};
-
-extern "C" void func_ov085_0212bedc(Obj *c)
-{
-    Vector3_local v;
-    Vector3_local res;
-    Vector3_local vd;
-
-    v.x = 0; v.y = 0; v.z = 0;
-    res.x = 0; res.y = 0; res.z = 0;
-    vd.x = 0; vd.y = 0; vd.z = 0;
-
-    vd.x = c->v5c;
-    vd.y = c->v60;
-    vd.z = c->v64;
-    vd.x = 0x1086000;
-    v.z = Vec3_HorzDist((Vector3_local *)&c->v5c, &vd);
-    Matrix4x3_FromRotationY(&data_020a0e68, Vec3_HorzAngle((Vector3_local *)&c->v5c, &vd));
-    MulVec3Mat4x3(&v, &data_020a0e68, &res);
-    {
-        int t;
-        vd.x = vd.x + res.x;
-        t = c->v60;
-        vd.y = t;
-        vd.z = vd.z + res.z;
-        Matrix4x3_FromTranslation(&data_020a0e68, vd.x >> 3, (t - 0xc000) >> 3, vd.z >> 3);
-    }
-    c->mtx = data_020a0e68;
-    _ZN8dActor_c19DropShadowRadHeightER11ShadowModelR9Matrix4x35Fix12IiES5_j((dActor_c *)c, *(ShadowModel *)&c->shadowmodel, *(Matrix4x3 *)&c->mtx, 0x46000, 0x258000, 0xf);
-}
